+++ Prime53/libs/Config.h
@@ -1,4 +1,4 @@
-extern uint MillerRabin_Lmt;
-extern uint MillerRabin_K;
+extern uint64 MillerRabin_Lmt;
+extern uint   MillerRabin_K;
 
 void LoadConfig(void);
